AzureKinectSensor class
