+++ src/Open3D/IO/Sensor/RGBDSensor.h
@@ -26,10 +26,24 @@
 
 #pragma once
 
+#include <memory>
+
+#include "Open3D/IO/Sensor/RGBDSensorConfig.h"
+
 namespace open3d {
+
+namespace geometry {
+class RGBDImage;
+};
+
 namespace io {
 
-class RGBDSensor {};
+class RGBDSensor {
+    RGBDSensor(const RGBDSensorConfig& rgbd_sensor_config);
+    virtual ~RGBDSensor(){};
+
+    virtual std::shared_ptr<geometry::RGBDImage> CaptureFrame() const = 0;
+};
 
 }  // namespace io
 }  // namespace open3d
